Added Account Page - Closes Issue 14
Added a basic account page, which can be viewed when the user is logged in. There the user can check their basic information and more sensitive information and as of now they can also edit their basic information. For that the necessary actions(EditAccount, ChangeBasicAccountSettings) were added, a series of viewmodels and the editAccount view.