Added Edit Email Feature - Closes Issue 17
With this update the user should be able to change the email address of their user account. For the implementation I just added a modal in the EditAccount view that appears when the user clicks on the button next to the email input and the actions RequestChangeEmail, EmailChangeVerificationMessage and EmailChangeVerificationMessage in the account controller. I also added the EmailChangeVerificationMessage view and all the necessary alert/messages for the UI.