Added DataAccess Library And Logic For The Posts Entity - Closes Issue 23
I added the dataaccess library, which can be found on the database schema of issue #22 . Also implemented the dataaccess logic for the Posts Entity and its sub entity PostVote. Finally added in the Authentication Library the entity friends as shown in the database schema of issue #22, this might be improved in the future.