Added DataAccess Logic For Chats Entity - Closes Issue 25
I added the basic implementation(CRUD) for the chats entity. Additionally I added the opportunity for a user to join a chat after it has been created and the option for a user to leave the chat without deleting it. Also added the ability for the posts to also have the post votes returned by their get methods(the join works correctly now).