Added Post FrontedEnd Implementation - Closes Issue 34
With this update a user is now able to view the recent posts. The user can also like and dislike those posts, which happens in real time and they can see without reload user liking and disliking posts. The user can also create posts and edit and delete their posts. Some minor changes were added in the PostDataAccess, which now has a method that allows the developer to specify how many posts they want to receive. Also SignalR support was added(and the PostVoteHub class) that allows for the real time communication between the clients.