Added Friendship System Front End Implementation - Closes Issue 36
I added a way for the user to be able to add and remove friends from their account page based on their username or email. As of now the users to do not have to accept or reject this "notification" since the system is not in place yet, but eventually there will be notifications, which the users will be able either accept or reject. Additionally it must be noted that I simplified the friendship entity since I could not manage to get the foreign keys to work in a many-to-many self referencing relationship, I might look into it more in future updates.