Implemented Front End For The Notification System - Closes Issue 38
This is a fairly large update, because the notification system influences a lot of parts of the Front End. More specifically in this update the notification controller was added with the view ViewNotifications and the hub NotificationHub. The balance between the controller(request response model) and the hub(SignalR) is a bit tenuous and I had to change parts of the layout for better code readability and other parts that had already implemented with the request response model(sending friend request in editAccount) to the SignalR architecture. There will be more additions to the system when the chat and messages front end are also implemented.