Implemented Chat Front End Interface - Closes Issue 40
After this update the users are able to create chats adding to them users from their friend list. The ability to send messages has not been implemented yet and it will be added in the next update. For the technical details, a Chat controller was added as was the ViewChats and ViewChat(empty as of yet) views. Also added js and css code for the multilect in site.css and site.js files. Fixed a minor bug in notification controller and changed a bit the ChatDataAccess to now return the users of the chat correctly.