Implemeted Message Front End Interface - Closes Issue 42
Added the ability for a user to send a receive messages in the chats they are part of. The messages are send asynchronously and received asynchronously from the other user if they are in the view of the specific chat that the message was sent for. If the a user of the chat is not connected at that moment a notification is sent to them asynchronously and their notification counter is updated asynchronously to notify them.  That means that the ViewNotifications view has also been updated and now the users can see their messages notifications there. Some minor changes have also been made for the models AppUser(now contains chatId field) and Message(now the connection it has with Notification Entity is a one-to-many). The message status model has also been added for the same reason.