Added Google External Identity Provider - Closes Issue 53
After this update a user can sign in the application using google as their external identity provider. If they have already an account they can still sign in using google and that external sign in is applied to their account. If then they wish to change account and they are using google as then external identity provider then that login provider is removed and if there is no password, because they created the local account using google as their external provider, then a new temp password is created, so that they can log in. Some minor bugs have been fixed and more emphasis was given in security(email credentials removed from appsettings.json and changed).