Added Microsoft External Identity Provider - Closes Issue 54
After this update a user can sign in the application using Microsoft as their external identity provider. If they have already an account they can still sign in using that account and then the external sign in is applied to that account seemlessly. The implementation is based mostly on the previous issue 53(add google as external identity provider) and thus only minor changes were made. Something that needs to be noted is that for some unknown reason only buisiness account are accepted through that method and personal hotmail accounts can not be used for this process. Hopefully this issue will be resolved in a future update.