Added Logging System - Closes Issue 60
I added a logging system that connects to the application insights system in azure. I made sure that that logging system only activates in production, but if necessary it can also be activated by simply changing the if statement in Program.cs to work with development. I also added multiple logs(information, warning, success) in the authentication library, dataaccess library and in the email sending library, where I thought they would be necessary. A minor bug was also fixed, which could lock the user out(using username-password login) if they a user created a local account and then continued with google or another external identity provider that connected to that account, before activating their local account with their confirmation email link.