Small External Identity Providers Change In Appsettings
I fixed a bug in which the application was crashing if the external authentication providers were not passed in the application as environmental variables or if they did not exist in the personal secrets.json. Now the application should not crash if the default values are not overriden through the file secrets.json or through environmental variables and only the external identity providers part of the application will not be funtional.